% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoEDA.R
\name{clean_ints}
\alias{clean_ints}
\title{Removes outliers based on quartiles or IQR}
\usage{
clean_ints(
  x,
  high = TRUE,
  low = TRUE,
  lower_q = 0.01,
  upper_q = 0.99,
  method = "quartile"
)
}
\arguments{
\item{x}{numeric variable}

\item{high}{clean upper bound}

\item{low}{clean lower bound}

\item{lower_q}{lower quartile}

\item{upper_q}{upper quartile}

\item{method}{method of cleaning either 'quartile' or 'iqr'}
}
\description{
Removes outliers based on quartiles or IQR
}
