% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoEDA.R
\name{create_eda_report}
\alias{create_eda_report}
\title{Export visualizations to powerpoint or pdf}
\usage{
create_eda_report(plots, path, format = "ppt")
}
\description{
Export visualizations to powerpoint or pdf
}
