% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoEDA.R
\name{viz_data}
\alias{viz_data}
\title{Visualize all features in dataset}
\usage{
viz_data(data, outcome, outcome_type, positive_class = NA)
}
\arguments{
\item{data}{dataframe}

\item{outcome}{outcome name in string format}

\item{outcome_type}{either 'classification' or 'regression'}

\item{positive_class}{if doing binary classification specify the positive class}
}
\description{
Visualize all features in dataset
}
