% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoEDA.R
\name{viz_factor}
\alias{viz_factor}
\title{Visualize categorical features}
\usage{
viz_factor(
  data,
  outcome,
  outcome_type,
  positive_class = NA,
  var,
  na_remove = T,
  lump = T,
  prop = 0.01,
  string_miss = T
)
}
\arguments{
\item{data}{dataframe}

\item{outcome}{outcome name in string format}

\item{outcome_type}{either 'classification' or 'regression'}

\item{positive_class}{if doing binary classification specify the positive class}

\item{var}{feature name in string format}

\item{na_remove}{specify the removal of outliers}

\item{lump}{group all levels underneath a certain threshold as 'other'}

\item{prop}{if using lump specify the proportion}
}
\description{
Visualize categorical features
}
