% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoEDA.R
\name{viz_int}
\alias{viz_int}
\title{Visualize numeric features}
\usage{
viz_int(
  data,
  var,
  outcome,
  outcome_type,
  positive_class = NA,
  remove_outliers = TRUE
)
}
\arguments{
\item{data}{dataframe}

\item{var}{feature name in string format}

\item{outcome}{outcome name in string format}

\item{outcome_type}{either 'regression' or 'classification'}

\item{positive_class}{if doing binary classification then specify the positive class}

\item{remove_outliers}{remove outliers via clean_ints}
}
\description{
Visualize numeric features
}
